Added confirmation dialog when deleting environments.

+++ frontend_app/app/js/controllers/indicator_controller.js
@@ -50,8 +50,10 @@ angular.module("app").controller("IndicatorController", function ($scope, $locat
   };
 
   $scope.destroyFlag = function(flag) {
-    FlagService.destroyFlag(flag);
-    $location.path("/flags");
+    if (window.confirm("Do you want to delete " + flag.title + " environment?")) {
+      FlagService.destroyFlag(flag);
+      $location.path("/flags");
+    }
   };
 });
